refactor: vinyldns
